[Feature]: Impl features -> Impl tontine bidding worker & repository

+++ src/main/kotlin/com/b2i/tontine/domain/tontine_periodicity/entity/TontinePeriodicity.kt
@@ -32,7 +32,7 @@ class TontinePeriodicity: BaseEntity() {
 
     var biddingDeadline : Date? = null
 
-    var biddingAmount : Long = 0
+    var biddingAmount : Double = 0.0
 
     @JsonBackReference
     @ManyToOne(fetch = FetchType.LAZY)
